Fix MGK, HMGN tests
