+++ test/models/meshgraphnet/test_meshgraphkan.py
@@ -20,29 +20,26 @@
 Unit tests for the MeshGraphKAN model.
 """
 
-import os
 import random
-import sys
 
 import numpy as np
 import pytest
 import torch
 
-script_path = os.path.abspath(__file__)
-sys.path.append(os.path.join(os.path.dirname(script_path), ".."))
+pytest.importorskip("torch_geometric")
 
 from test import common  # noqa: E402
 from test.conftest import requires_module  # noqa: E402
+from test.models.meshgraphnet.utils import rand_graph
 
-dgl = pytest.importorskip("dgl")
 
-
-@requires_module("dgl")
+@requires_module("torch_geometric")
 def test_meshgraphkan_forward(device, pytestconfig, set_physicsnemo_force_te):
+    import torch_geometric as pyg
+
     from physicsnemo.models.meshgraphnet import MeshGraphKAN
 
     torch.manual_seed(0)
-    dgl.seed(0)
     np.random.seed(0)
 
     model = MeshGraphKAN(4, 3, 2).to(device)
@@ -52,11 +49,15 @@ def test_meshgraphkan_forward(device, pytestconfig, set_physicsnemo_force_te):
     for _ in range(bsize):
         src = torch.tensor([np.random.randint(n_nodes) for _ in range(n_edges)])
         dst = torch.tensor([np.random.randint(n_nodes) for _ in range(n_edges)])
-        graphs.append(dgl.graph((src, dst)).to(device))
-    graph = dgl.batch(graphs)
+        graphs.append(
+            pyg.data.Data(
+                edge_index=torch.stack([src, dst], dim=0), num_nodes=n_nodes
+            ).to(device)
+        )
+    graph = pyg.data.Batch.from_data_list(graphs)
 
-    node_f = torch.randn(graph.num_nodes(), 4).to(device)
-    edge_f = torch.randn(graph.num_edges(), 3).to(device)
+    node_f = torch.randn(graph.num_nodes, 4).to(device)
+    edge_f = torch.randn(graph.num_edges, 3).to(device)
 
     assert common.validate_forward_accuracy(
         model,
@@ -65,8 +66,10 @@ def test_meshgraphkan_forward(device, pytestconfig, set_physicsnemo_force_te):
     )
 
 
-@requires_module("dgl")
+@requires_module("torch_geometric")
 def test_meshgraphkan_constructor(device, pytestconfig, set_physicsnemo_force_te):
+    import torch_geometric as pyg
+
     arg_sets = [
         dict(
             input_dim_nodes=random.randint(2, 8),
@@ -100,28 +103,30 @@ def test_meshgraphkan_constructor(device, pytestconfig, set_physicsnemo_force_te
         model = MeshGraphKAN(**kw).to(device)
         bsize = random.randint(1, 4)
         n_nodes, n_edges = random.randint(8, 15), random.randint(8, 15)
-        graph = dgl.batch(
-            [dgl.rand_graph(n_nodes, n_edges).to(device) for _ in range(bsize)]
+        graph = pyg.data.Batch.from_data_list(
+            [rand_graph(n_nodes, n_edges, device) for _ in range(bsize)]
         )
-        node_f = torch.randn(graph.num_nodes(), kw["input_dim_nodes"]).to(device)
-        edge_f = torch.randn(graph.num_edges(), kw["input_dim_edges"]).to(device)
+        node_f = torch.randn(graph.num_nodes, kw["input_dim_nodes"]).to(device)
+        edge_f = torch.randn(graph.num_edges, kw["input_dim_edges"]).to(device)
         out = model(node_f, edge_f, graph)
-        assert out.shape == (graph.num_nodes(), kw["output_dim"])
+        assert out.shape == (graph.num_nodes, kw["output_dim"])
 
 
-@requires_module("dgl")
+@requires_module("torch_geometric")
 def test_meshgraphkan_optims(device, pytestconfig, set_physicsnemo_force_te):
+    import torch_geometric as pyg
+
     from physicsnemo.models.meshgraphnet import MeshGraphKAN
 
     def make_inputs():
         model = MeshGraphKAN(3, 3, 2).to(device)
         bsize = random.randint(1, 6)
         n_nodes, n_edges = random.randint(10, 20), random.randint(10, 20)
-        graph = dgl.batch(
-            [dgl.rand_graph(n_nodes, n_edges).to(device) for _ in range(bsize)]
+        graph = pyg.data.Batch.from_data_list(
+            [rand_graph(n_nodes, n_edges, device) for _ in range(bsize)]
         )
-        node_f = torch.randn(graph.num_nodes(), 3).to(device)
-        edge_f = torch.randn(graph.num_edges(), 3).to(device)
+        node_f = torch.randn(graph.num_nodes, 3).to(device)
+        edge_f = torch.randn(graph.num_edges, 3).to(device)
         return model, [node_f, edge_f, graph]
 
     m, inp = make_inputs()
@@ -134,27 +139,27 @@ def test_meshgraphkan_optims(device, pytestconfig, set_physicsnemo_force_te):
     assert common.validate_combo_optims(m, (*inp,))
 
 
-@requires_module("dgl")
+@requires_module("torch_geometric")
 def test_meshgraphkan_checkpoint(device, pytestconfig, set_physicsnemo_force_te):
     from physicsnemo.models.meshgraphnet import MeshGraphKAN
 
     m1 = MeshGraphKAN(4, 3, 4).to(device)
     m2 = MeshGraphKAN(4, 3, 4).to(device)
 
-    graph = dgl.rand_graph(12, 18).to(device)
+    graph = rand_graph(12, 18, device)
     node_f = torch.randn(12, 4).to(device)
     edge_f = torch.randn(18, 3).to(device)
 
     assert common.validate_checkpoint(m1, m2, (node_f, edge_f, graph))
 
 
-@requires_module("dgl")
+@requires_module("torch_geometric")
 @common.check_ort_version()
 def test_meshgraphkan_deploy(device, pytestconfig, set_physicsnemo_force_te):
     from physicsnemo.models.meshgraphnet import MeshGraphKAN
 
     model = MeshGraphKAN(5, 2, 3).to(device)
-    graph = dgl.rand_graph(14, 25).to(device)
+    graph = rand_graph(14, 25, device)
     node_f = torch.randn(14, 5).to(device)
     edge_f = torch.randn(25, 2).to(device)
     inputs = (node_f, edge_f, graph)
